#35 Split model building into a ModelBuilder
Move the SBML-to-internal-representation building out of ChasteSbmlModel into a
ModelBuilder class: all the _format_* translators, _add_* data builders, ODE
extraction, equation ordering and the name-conflict check, along with the
collections they populate. ChasteSbmlModel becomes a thin orchestrator -- load
the SBML, resolve id conflicts, build, populate the template context and render --
reading the built collections through ModelBuilder.template_data().

Behaviour-neutral: reference model output is byte-identical. The _check_name_conflicts
tests move to test_model_builder.py, and a new test_chaste_sbml_model.py covers the
orchestrator (write and the outputs property) end-to-end.
